moved signing keys to app so it can be shared across views.
