+++ src/features/router/Router.tsx
@@ -1,3 +1,5 @@
+import { FC } from "react";
+
 import { BrowserRouter, Switch, Route } from "react-router-dom";
 
 // Pages
@@ -7,18 +9,23 @@ import Vote from "features/vote";
 import Navbar from "common/components/navbar";
 import Wallet from "features/wallet";
 import Footer from "common/components/footer";
+import { SigningKeys } from "App";
 
-const Router = () => {
+interface Props {
+  signingKeys: SigningKeys;
+}
+
+const Router: FC<Props> = ({ signingKeys }) => {
   return (
     <BrowserRouter>
       <div>
         <Navbar />
         {/* A <Switch> looks through its children <Route>s and
             renders the first one that matches the current URL. */}
-        <Wallet />
+        <Wallet signingKeys={signingKeys} />
         <Switch>
           <Route path="/">
-            <Vote />
+            <Vote signingKeys={signingKeys} />
           </Route>
         </Switch>
       </div>
